add primitive music
this isn't a great use of time, if I want music I should use a library
to do the playback and use a better notation to write it.
The thing I wanted custom was playback so parts could be removed when
things went faster but existing instruments and notation systems can be
used for this.